+++ src/main/php/lang/iterator/MappingIterator.php
@@ -10,7 +10,7 @@
 namespace stubbles\lang\iterator;
 use stubbles\lang;
 /**
- * Maps values and optionally keys from an underlying iterator.
+ * Maps values and/or keys from an underlying iterator.
  *
  * @since  5.0.0
  */
@@ -32,15 +32,20 @@ class MappingIterator extends \IteratorIterator
     /**
      * constructor
      *
-     * @param  \Traversable  $iterator     iterator to map values of
-     * @param  callable      $valueMapper  callable which maps the values
-     * @param  callable      ‚$keyMapper    callable which maps the keys
+     * @param   \Traversable  $iterator     iterator to map values of
+     * @param   callable      $valueMapper  optional  callable which maps the values
+     * @param   callable      $keyMapper    optional  callable which maps the keys
+     * @throws  \InvalidArgumentException  in case both $valueMapper and $keyMapper are null
      */
-    public function __construct(\Traversable $iterator, callable $valueMapper, callable $keyMapper = null)
+    public function __construct(\Traversable $iterator, callable $valueMapper = null, callable $keyMapper = null)
     {
+        if (null === $valueMapper && null === $keyMapper) {
+            throw new \InvalidArgumentException('Passed null for both valueMapper and keyMapper, but at least one of both must not be null');
+        }
+
         parent::__construct($iterator);
-        $this->valueMapper = lang\ensureCallable($valueMapper);
-        $this->keyMapper   = null !== $keyMapper ? lang\ensureCallable($keyMapper) : (null);
+        $this->valueMapper = null !== $valueMapper ? lang\ensureCallable($valueMapper) : null;
+        $this->keyMapper   = null !== $keyMapper ? lang\ensureCallable($keyMapper) : null;
     }
 
     /**
@@ -50,6 +55,10 @@ class MappingIterator extends \IteratorIterator
      */
     public function current()
     {
+        if (null === $this->valueMapper) {
+            return parent::current();
+        }
+
         $map = $this->valueMapper;
         return $map(parent::current(), parent::key());
     }
@@ -66,6 +75,6 @@ class MappingIterator extends \IteratorIterator
         }
 
         $map = $this->keyMapper;
-        return $map(parent::key());
+        return $map(parent::key(), parent::current());
     }
 }